+++ src/org/eclipse/swt/internal/image/PngDecodingDataStream.d
@@ -95,9 +95,9 @@ public override int read() {
     return nextDecodedByte & 0xFF;
 }
 
-public override int read(byte[] buffer, int off, int len) {
-    for (int i = 0; i < len; i++) {
-        int b = read();
+public override ptrdiff_t read(byte[] buffer, ptrdiff_t off, ptrdiff_t len) {
+    for (size_t i = 0; i < len; i++) {
+        size_t b = read();
         if (b is -1) return i;
         buffer[off + i] = cast(byte)b;
     }
