+++ src/org/eclipse/swt/widgets/FontDialog.d
@@ -170,7 +170,7 @@ public FontData open () {
         display.setModalDialog (this);
     }
     int signalId = 0;
-    int /*long*/ hookId = 0;
+    ptrdiff_t hookId = 0;
     CallbackData emissionData;
     emissionData.display = display;
     emissionData.data = handle;
