+++ src/org/eclipse/swt/widgets/Menu.d
@@ -498,7 +498,7 @@ public MenuItem [] getItems () {
 override String getNameText () {
     String result = "";
     MenuItem [] items = getItems ();
-    int length_ = items.length;
+    ptrdiff_t length_ = items.length;
     if (length_ > 0) {
         for (int i=0; i<length_-1; i++) {
             result = result ~ items [i].getNameText() ~ ", ";
@@ -608,7 +608,7 @@ public bool getVisible () {
     return OS.GTK_WIDGET_MAPPED (handle);
 }
 
-override int /*long*/ gtk_hide (GtkWidget* widget) {
+override int gtk_hide (GtkWidget* widget) {
     if ((style & SWT.POP_UP) !is 0) {
         display.activeShell = getShell ();
     }
@@ -627,7 +627,7 @@ override int /*long*/ gtk_hide (GtkWidget* widget) {
     return 0;
 }
 
-override int /*long*/ gtk_show (GtkWidget* widget) {
+override int gtk_show (GtkWidget* widget) {
     if ((style & SWT.POP_UP) !is 0) {
         display.activeShell = getShell ();
         return 0;
@@ -637,7 +637,7 @@ override int /*long*/ gtk_show (GtkWidget* widget) {
 }
 
 
-override int /*long*/ gtk_show_help (GtkWidget* widget, int /*long*/ helpType) {
+override int gtk_show_help (GtkWidget* widget, ptrdiff_t helpType) {
     if (sendHelpEvent (helpType)) {
         OS.gtk_menu_shell_deactivate (cast(GtkMenuShell*)handle);
         return 1;
@@ -843,7 +843,7 @@ public void removeHelpListener (HelpListener listener) {
     eventTable.unhook (SWT.Help, listener);
 }
 
-bool sendHelpEvent (int /*long*/ helpType) {
+bool sendHelpEvent (ptrdiff_t helpType) {
     if (selectedItem !is null && !selectedItem.isDisposed()) {
         if (selectedItem.hooks (SWT.Help)) {
             selectedItem.postEvent (SWT.Help);
