+++ src/org/eclipse/swt/widgets/MessageBox.d
@@ -184,7 +184,7 @@ public int open () {
         display.setModalDialog (this);
     }
     int signalId = 0;
-    int /*long*/ hookId = 0;
+    ptrdiff_t hookId = 0;
     CallbackData emissionData;
     emissionData.display = display;
     emissionData.data = handle;
