+++ src/org/eclipse/swt/widgets/Tracker.d
@@ -449,7 +449,7 @@ override int gtk_key_press_event (GtkWidget* widget, GdkEventKey* keyEvent) {
              * values have changed.
              */
             if (rectangles !is oldRectangles) {
-                int length = rectangles.length;
+                ptrdiff_t length = rectangles.length;
                 if (length !is rectsToErase.length) {
                     draw = true;
                 } else {
@@ -494,7 +494,7 @@ override int gtk_key_press_event (GtkWidget* widget, GdkEventKey* keyEvent) {
              * values have changed.
              */
             if (rectangles !is oldRectangles) {
-                int length = rectangles.length;
+                ptrdiff_t length = rectangles.length;
                 if (length !is rectsToErase.length) {
                     draw = true;
                 } else {
@@ -573,7 +573,7 @@ private int gtk_mouse (int eventType, GtkWidget* widget, void* eventPtr) {
              * values have changed.
              */
             if (rectangles !is oldRectangles) {
-                int length = rectangles.length;
+                ptrdiff_t length = rectangles.length;
                 if (length !is rectsToErase.length) {
                     draw = true;
                 } else {
@@ -618,7 +618,7 @@ private int gtk_mouse (int eventType, GtkWidget* widget, void* eventPtr) {
              * values have changed.
              */
             if (rectangles !is oldRectangles) {
-                int length = rectangles.length;
+                ptrdiff_t length = rectangles.length;
                 if (length !is rectsToErase.length) {
                     draw = true;
                 } else {
@@ -970,7 +970,7 @@ public void setRectangles (Rectangle [] rectangles) {
     checkWidget();
     // SWT extension: allow null for zero length string
     //if (rectangles is null) error (SWT.ERROR_NULL_ARGUMENT);
-    int length = rectangles.length;
+    ptrdiff_t length = rectangles.length;
     this.rectangles = new Rectangle [length];
     for (int i = 0; i < length; i++) {
         Rectangle current = rectangles [i];
